Split general settings into MVC files
